Updated cmdline code
